- [file-manipulator] changing cursor is not working.
by the standard docs, changing cursor for writing something on "a+" opened file is prohibited.
Actually, it goes back to the end of file before write something automatically.
But, "w+" is not useful, because it always overwrites existing file.

So for users who use this API for using appending something but if it exists already just append, we ready "APPENDABLE" mode.