- [wrd] change a convention rule for prefix on retrieving state flag.
we used to add "does" for quering status for some flag, but for now,
"is" instead.
adding "is" looks weird for some func name, for instance isOccupy.
but I thing it's more valuable for putting same prefix for similar func name.
it makes user can anticipate what it can do easily when they saw it added "is".