+ [wrd] change WRD_GET to do another action.
WRD_GET was used to check given arg was null obj of WRD frx and
return immediately if so.
it reduces standard c++ null-check boiler-plate codes, most of them
are 3 lines per case, into 1 single line. but it still makes more lines
when user tries to access obj inside of another obj, such as anti demeter's law.

WRD_GET now can be likely to C#'s null propagation syntax.
(e.g, as-is:
 A& a = getA();
 B& b = a.getB();
 C& c = b.getC();

 to-be:
 C& c = WRD_GET(getA(), getB(), getC()); // it's pretty intuitive if you can
                                         // see those commas as dots.

but this may have some disadvantages and need for us to care when you use it.