interp: refactor: restructure 'pack' creation
in order to manage HANDLE of native library loaded, a pack instance
needs to hold a reference of it if it's from native.

so we can't use single pack class. we won't allow that all pack
instances must take a reference even though most of them don't need it.

so pack class will be various in future. that't why I made 'packMaker'.
now, with packMaker, a pack will be made by it for the given situation.
no matter what the pack is, native or managed or from java file, system
won't care and just calls subs() function to let them prepare origin
objects from their sources.