interp: refactor: parser should generate AST.
it should be treated as singlethread.
only job it has is control lowparser and lowscanner, they are generated
parser and scanner from like bison, and give some API to interact to
other components as facade.

components which try to get some AST should use parser and verifier on
their own to generate valid AST tree.
during that operation, using multithreaded envrionment is up to those
components.