interp: feat: pack is able to have src info
'src' in this context means that srcFile to be refered by expression.
when process unhandled exception object the interpreter needs to show
logs callstack. the callstack should contains info which file and
linenumber had been executed.

exceptions only occurs when run 'expr' object.

so it's natural for only expr instances to hold line number of source
code and refer to src object which get filename of source code.