wrd: refactor: improve implementation of immutable
previously I've implemented immutable feature with scope class.
scope class generates and wraps given elements when user want to add
something. and the wrapper judge to clone a new object or not when
it's going to bind something.

however, in this scenario user can have nested ref object which was
refering to another ref. so when user tries to cast from node class to
specific type derived from node, F/C may occurs.

e.g.
    node* a = new ref(new ref(new myObj()));
    node* b = new myObj();

    b->cast<myObj>(); // is success.
    a->cast<myObj>(); // is failure.
    a->cast<ref>().get().cast<ref>().get().cast<myObj>(); // is success.

and this is so silly.

to overcome this limitation, I removed ref and let it be derived from
str class.
now container using str to manages its elements can guarrantee that
all of its elements are not refer.