wrd: refactor: remove common container
map is not applicable to be handled like a container.
it needs 2 terms to operate funcs.
of course, I can make a data class holding those 2 terms and let a
map take it as a its type parameter.

but if I make it be possible, it is nearly immpossible to define a
generic func which works well on both of an array and a map.
so it is valueless to define a common container.