wrd: refactor: remove doesNeedScope()
all objects including native one are needed to be make a frame
instance before run sub nodes.

I thought a possibility of not putting 'me' object as an argument
when run sub nodes. however, in this case we need to distinguish
the sub node which we try to run is a static func or instance func.
wrd lang normally doesn't have any concept of static func.
the most funcs are instance func. except for constructor.

so by considering all of this requirments I decided to treat all
funcs as same way. so I'll make a frame instance when native call.
even though the frame instance won't be used inside of the call.