wrd: fix: evalType of func is the func, not return type
please don't confuse this to evalType.
evaluate means program just read the code and expand or calculate then
replace it to the accurate another identifier or variable.
so evalType of the func is just the func itself because there is nothing
to calculate or run.
that's completely different to Return type of the function.
if there is a runExpr instance, and it takes the func and proper argument,
then it can be evaluated and its evalType is the return type of the func.