wrd: refactor: ctor placed as a func
before, ctor had treated when user writes code 'objName()' so I let it
be called when run() of obj got called.

in conclusion, that was mistake.

because sometimes I need to make an AST structure to return obj instance
itself with run().

this commit changes those ctors are managed as a funcs.
so each object will get '@ctor' funcs. and in order to call them,
AST structure need to be made to call run("@ctor", obj)
when user codes like "objName()"