core: fix: remove ctor at point class
if there is ctor on struct and when compiler compiles it on c like codes,
it recognizes point class that it's like a cpp class, not c class.

c struct and cpp struct are quite different when compilers tries to make
an instance, as you may know.
so that's the problem.