core: refactor: redefine 'ctor' class
Traditionally, calling ctor() would create an object inside.
But with the advent of @common, sometimes you need to call two constructors.
So the object creation is only done inside baseObj::run(),
and the constructors are supposed to act as simple functions instead of creating objects directly.
So, naturally, we wanted to change the Ctor class to a func.

The problem was that when created as a bridgeCtor, the bridge object could not hold the
meta-information that would require a constructor call with arguments.
In summary, it is natural for bridge objects to create objects via bridgeCtor, and for mgdObjs
to create objects via baseObj, and for constructors to be functions.

To merge these two structures as naturally as possible, we implemented the following
The ctor class (mgdObj's constructor) always returns an obj as its return value, which is
contained in args.

baseObj will clone the object as before. Then we can call @common and return the result of the
call to @ctor directly.

bridgeCtor will create and return an object as before.

However, there are two caveats.
1) You should not provide a syntax that allows you to call only the constructor directly from
managed code.

In this case, you might want to call only the bridge object's constructor.
(Note that managed cannot and should not distinguish whether this is a bridge object or a managed object).
you would only want to call the constructor of an object that already exists,
but the object is actually created.

2) You should consider what happens if you inherit from a bridge object.

When bridgeCtor is called, a bridge object is created that is separate from the inherited
managed object.